Update for Processing 2.0b8

+++ src/peasy/PeasyWheelHandler.java
@@ -1,5 +1,5 @@
 package peasy;
 
 public interface PeasyWheelHandler {
-	public void handleWheel(final int delta);
+	public void handleWheel(final float delta);
 }
